🌐 domain: rename challenge star error contract
