+++ packages/core/src/challenging/domain/errors/index.ts
@@ -6,5 +6,5 @@ export { NotSolutionAuthorError } from './NotSolutionAuthorError'
 export { SolutionTitleAlreadyInUseError } from './SolutionTitleAlreadyInUseError'
 export { ChallengeAlreadyExistsError } from './ChallengeAlreadyExistsError'
 export { ChallengeProblemNotFoundError } from './ChallengeProblemNotFoundError'
-export { ChallengeIsAlreadyStarError } from './ChallengeIsAleadyStarError'
+export { ChallengeIsAlreadyStarError } from './ChallengeIsAlreadyStarError'
 export { ChallengeIsNotStarChallengeError } from './ChallengeIsNotStarChallengeError'
